requirements.txt
black==24.3.0
Django==4.2.10
django-upgrade==1.15.0
djangorestframework==3.15.1
flake8==6.1.0
flake8-django==1.4
gunicorn==22.0.0
isort==5.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-treenode==0.22.0